pyproject.toml
[project]
name = "playwrightcicd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-html>=4.2.0",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.2",
]

[tool.pytest.ini_options]
markers = [
    "api: API tests",
    "ui: UI tests"
]
